// packages/octez.connect-blockchain-substrate/src/types/messages/permission-response.ts
import { AppMetadata, PermissionResponseV3 } from '@tezos-x/octez.connect-types'
import { SubstratePermissionScope } from '../permission-scope'

export interface SubstratePermissionResponse extends PermissionResponseV3<'substrate'> {
  blockchainData: {
    appMetadata: AppMetadata
    scopes: SubstratePermissionScope[] // enum
    accounts: {
      accountId: string
      network?: {
        genesisHash: string
        rpc?: string
      }
      publicKey: string
      address: string
    }[]
  }
}